Rework client usage.
