+++ packages/nimbus/scripts/AutoRelease.js
@@ -1,6 +1,5 @@
 const { Script } = require('@beemo/core');
-const { createGitHubClient, parseGitRepo } = require('@lerna/github-client');
-const { getLastTag, getCommitsSince } = require('@airbnb/nimbus-common/git');
+const createGitHubClient = require('./helpers/createGitHubClient');
 const { LERNA_VERSION_ARGS } = require('./constants');
 
 // Primarily used within CI jobs
@@ -11,10 +10,8 @@ module.exports = class AutoReleaseScript extends Script {
 
   bootstrap() {
     this.task('Setting git environment variables', this.setGitEnvVars);
-    this.task('Getting git commits since last tag', this.getCommitsSinceLastTag);
     this.task('Bumping package versions', this.versionPackages);
     this.task('Publishing packages to NPM', this.publishPackages);
-    this.task('Adding label to GitHub PRs', this.addLabelToPRs);
   }
 
   // https://git-scm.com/book/en/v2/Git-Internals-Environment-Variables
@@ -50,37 +47,7 @@ module.exports = class AutoReleaseScript extends Script {
       GIT_TERMINAL_PROMPT: 0,
     });
 
-    // Required by Lerna
-    if (env.GITHUB_TOKEN) {
-      process.env.GH_TOKEN = env.GITHUB_TOKEN;
-    }
-
     context.client = createGitHubClient();
-    context.repo = parseGitRepo();
-  }
-
-  // Commit looks like:
-  // 2c595c5 (HEAD -> refs/heads/master, refs/remotes/origin/master, refs/remotes/origin/HEAD) Docs: Update readme.
-  getCommitsSinceLastTag(context) {
-    context.commits = [];
-    context.prs = [];
-
-    return getLastTag().then(tag => {
-      context.lastTag = tag;
-
-      return getCommitsSince(tag).then(commits => {
-        commits.forEach(commit => {
-          context.commits.push(commit.trim());
-
-          // Extract the PR number if available
-          const pr = commit.match(/\(#(\d+)\)/);
-
-          if (pr && pr[1]) {
-            context.prs.push(Number(pr[1]));
-          }
-        });
-      });
-    });
   }
 
   // https://github.com/lerna/lerna/tree/master/commands/version#readme
@@ -98,20 +65,6 @@ module.exports = class AutoReleaseScript extends Script {
         // Run pre and post scripts in each package if available
         '--require-scripts',
       ]),
-    );
-  }
-
-  // https://octokit.github.io/rest.js/#api-Issues-addLabels
-  addLabelToPRs(context) {
-    return Promise.all(
-      context.prs.map(number =>
-        context.client.request('POST /repos/:owner/:repo/issues/:number/labels', {
-          owner: context.repo.owner,
-          repo: context.repo.name,
-          number,
-          data: ['released'],
-        }),
-      ),
     );
   }
