+++ packages/nimbus/scripts/helpers/createGitHubClient.js
@@ -1,0 +1,24 @@
+const Octokit = require('@octokit/rest');
+const pkg = require('../../package.json');
+
+module.exports = function createGitHubClient() {
+  const { GITHUB_TOKEN, GHE_API_URL, GHE_VERSION } = process.env;
+  const options = {
+    userAgent: `Nimbus v${pkg.version}`,
+  };
+
+  if (GITHUB_TOKEN) {
+    options.auth = `token ${GITHUB_TOKEN}`;
+  }
+
+  if (GHE_API_URL) {
+    options.baseUrl = GHE_API_URL;
+  }
+
+  if (GHE_VERSION) {
+    // eslint-disable-next-line
+    Octokit.plugin(require(`@octokit/plugin-enterprise-rest/ghe-${GHE_VERSION}`));
+  }
+
+  return new Octokit(options);
+};
